ValhallaAttributeTests
- move attributes to ValhallaUtils
- remove testValueTypeClassWithImplicitCreationAttribute since there are now uses of ImplicitAttribute in ValueTypeTests
