+++ test/functional/Valhalla/src/org/openj9/test/lworld/ValhallaAttributeGenerator.java
@@ -30,61 +30,55 @@ public class ValhallaAttributeGenerator extends ClassLoader {
 
 	public static Class<?> generateClassWithTwoPreloadAttributes(String name, String[] classList1, String[] classList2) throws Throwable {
 		byte[] bytes = generateClass(name, ValhallaUtils.ACC_IDENTITY, new Attribute[] {
-			new PreloadAttribute(classList1),
-			new PreloadAttribute(classList2)});
+			new ValhallaUtils.PreloadAttribute(classList1),
+			new ValhallaUtils.PreloadAttribute(classList2)});
 		return generator.defineClass(name, bytes, 0, bytes.length);
 	}
 
 	public static Class<?> generateClassWithPreloadAttribute(String name, String[] classList) throws Throwable {
-		byte[] bytes = generateClass(name, ValhallaUtils.ACC_IDENTITY, new Attribute[] {new PreloadAttribute(classList)});
+		byte[] bytes = generateClass(name, ValhallaUtils.ACC_IDENTITY, new Attribute[] {new ValhallaUtils.PreloadAttribute(classList)});
 		return generator.defineClass(name, bytes, 0, bytes.length);
 	}
 
 	public static Class<?> generateClassWithTwoImplicitCreationAttributes(String name) throws Throwable {
 		byte[] bytes = generateClass(name, ACC_FINAL + ValhallaUtils.ACC_VALUE_TYPE,
-			new Attribute[] {new ImplicitCreationAttribute(), new ImplicitCreationAttribute()});
+			new Attribute[] {new ValhallaUtils.ImplicitCreationAttribute(), new ValhallaUtils.ImplicitCreationAttribute()});
 		return generator.defineClass(name, bytes, 0, bytes.length);
 	}
 
 	public static Class<?> generateNonValueTypeClassWithImplicitCreationAttribute(String name) throws Throwable {
-		byte[] bytes = generateClass(name, ValhallaUtils.ACC_IDENTITY, new Attribute[] {new ImplicitCreationAttribute()});
-		return generator.defineClass(name, bytes, 0, bytes.length);
-	}
-
-	public static Class<?> generateValidClassWithImplicitCreationAttribute(String name) throws Throwable {
-		byte[] bytes = generateClass(name, ACC_FINAL + ValhallaUtils.ACC_VALUE_TYPE,
-			new Attribute[] {new ImplicitCreationAttribute()});
+		byte[] bytes = generateClass(name, ValhallaUtils.ACC_IDENTITY, new Attribute[] {new ValhallaUtils.ImplicitCreationAttribute()});
 		return generator.defineClass(name, bytes, 0, bytes.length);
 	}
 
 	public static Class<?> generateFieldWithMultipleNullRestrictedAttributes(String className, String fieldClassName) throws Throwable {
 		/* Generate field class - value class with ImplicitCreation attribute and ACC_DEFAULT flag set  */
 		byte[] fieldClassBytes = generateClass(fieldClassName, ACC_FINAL + ValhallaUtils.ACC_VALUE_TYPE,
-			new Attribute[] {new ImplicitCreationAttribute()});
+			new Attribute[] {new ValhallaUtils.ImplicitCreationAttribute()});
 		Class<?> fieldClass = generator.defineClass(fieldClassName, fieldClassBytes, 0, fieldClassBytes.length);
 
 		/* Generate class with field and multiple NullRestricted attributes */
 		byte[] classBytes = generateIdentityClassWithField(className, 0,
-			"field", fieldClass.descriptorString(), new Attribute[]{new NullRestrictedAttribute(), new NullRestrictedAttribute()});
+			"field", fieldClass.descriptorString(), new Attribute[]{new ValhallaUtils.NullRestrictedAttribute(), new ValhallaUtils.NullRestrictedAttribute()});
 		return generator.defineClass(className, classBytes, 0, classBytes.length);
 	}
 
 	public static Class<?> generateNullRestrictedAttributeInPrimitiveField(String className) throws Throwable {
 		/* Generate class with primitive field and a NullRestricted attribute */
 		byte[] classBytes = generateIdentityClassWithField(className, 0,
-			"field", "I", new Attribute[]{new NullRestrictedAttribute()});
+			"field", "I", new Attribute[]{new ValhallaUtils.NullRestrictedAttribute()});
 		return generator.defineClass(className, classBytes, 0, classBytes.length);
 	}
 
 	public static Class<?> generateNullRestrictedAttributeInArrayField(String className, String fieldClassName) throws Throwable {
 		/* Generate field class - value class with ImplicitCreation attribute and ACC_DEFAULT flag set.  */
 		byte[] fieldClassBytes = generateClass(fieldClassName, ACC_FINAL + ValhallaUtils.ACC_VALUE_TYPE,
-			new Attribute[] {new ImplicitCreationAttribute()});
+			new Attribute[] {new ValhallaUtils.ImplicitCreationAttribute()});
 		Class<?> fieldClass = generator.defineClass(fieldClassName, fieldClassBytes, 0, fieldClassBytes.length);
 
 		/* Generate class field field that is an array with a NullRestricted attribute */
 		byte[] classBytes = generateIdentityClassWithField(className, 0,
-			"field", "[" + fieldClass.descriptorString(), new Attribute[]{new NullRestrictedAttribute()});
+			"field", "[" + fieldClass.descriptorString(), new Attribute[]{new ValhallaUtils.NullRestrictedAttribute()});
 		return generator.defineClass(className, classBytes, 0, classBytes.length);
 	}
 
@@ -93,7 +87,7 @@ public class ValhallaAttributeGenerator extends ClassLoader {
 
 		/* Generate field class - value class with ImplicitCreation attribute and ACC_DEFAULT flag set.  */
 		byte[] fieldClassBytes = generateClass(fieldClassName, ACC_PUBLIC + ACC_FINAL + ValhallaUtils.ACC_VALUE_TYPE,
-			new Attribute[] {new ImplicitCreationAttribute()});
+			new Attribute[] {new ValhallaUtils.ImplicitCreationAttribute()});
 		Class<?> fieldClass = generator.defineClass(fieldClassName, fieldClassBytes, 0, fieldClassBytes.length);
 
 		ClassWriter classWriter = new ClassWriter(0);
@@ -101,7 +95,7 @@ public class ValhallaAttributeGenerator extends ClassLoader {
 
 		/* static field of previously generated field class with NullRestrictd attribute */
 		FieldVisitor fieldVisitor = classWriter.visitField(ACC_PUBLIC + ACC_STATIC, fieldName, fieldClass.descriptorString(), null, null);
-		fieldVisitor.visitAttribute(new NullRestrictedAttribute());
+		fieldVisitor.visitAttribute(new ValhallaUtils.NullRestrictedAttribute());
 
 		/* assign field to null in <clinit> */
 		MethodVisitor mvClinit = classWriter.visitMethod(ACC_STATIC, "<clinit>", "()V", null, null);
@@ -131,7 +125,7 @@ public class ValhallaAttributeGenerator extends ClassLoader {
 
 		/* Generate field class - value class with ImplicitCreation attribute and ACC_DEFAULT flag set.  */
 		byte[] fieldClassBytes = generateClass(fieldClassName, ACC_PUBLIC + ACC_FINAL + ValhallaUtils.ACC_VALUE_TYPE,
-			new Attribute[] {new ImplicitCreationAttribute()});
+			new Attribute[] {new ValhallaUtils.ImplicitCreationAttribute()});
 		Class<?> fieldClass = generator.defineClass(fieldClassName, fieldClassBytes, 0, fieldClassBytes.length);
 
 		ClassWriter classWriter = new ClassWriter(0);
@@ -139,7 +133,7 @@ public class ValhallaAttributeGenerator extends ClassLoader {
 
 		/* instance field of previously generated field class with NullRestrictd attribute */
 		FieldVisitor fieldVisitor = classWriter.visitField(ACC_PUBLIC, fieldName, fieldClass.descriptorString(), null, null);
-		fieldVisitor.visitAttribute(new NullRestrictedAttribute());
+		fieldVisitor.visitAttribute(new ValhallaUtils.NullRestrictedAttribute());
 
 		/* assign field to null in <init> */
 		MethodVisitor mvInit = classWriter.visitMethod(ACC_PUBLIC, "<init>", "()V", null, null);
@@ -165,7 +159,7 @@ public class ValhallaAttributeGenerator extends ClassLoader {
 
 		/* Generate class with field that is an identity class with a NullRestricted attribute */
 		byte[] classBytes = generateIdentityClassWithField(className, isStatic ? ACC_STATIC : 0,
-			"field", fieldClass.descriptorString(), new Attribute[]{new NullRestrictedAttribute()});
+			"field", fieldClass.descriptorString(), new Attribute[]{new ValhallaUtils.NullRestrictedAttribute()});
 		return generator.defineClass(className, classBytes, 0, classBytes.length);
 	}
 
@@ -176,19 +170,19 @@ public class ValhallaAttributeGenerator extends ClassLoader {
 
 		/* Generate class with field that has a NullRestricted attribute */
 		byte[] classBytes = generateIdentityClassWithField(className, isStatic ? ACC_STATIC : 0,
-			"field", fieldClass.descriptorString(), new Attribute[]{new NullRestrictedAttribute()});
+			"field", fieldClass.descriptorString(), new Attribute[]{new ValhallaUtils.NullRestrictedAttribute()});
 		return generator.defineClass(className, classBytes, 0, classBytes.length);
 	}
 
 	public static Class<?> generateNullRestrictedFieldWhereICHasNoDefaultFlag(boolean isStatic, String className, String fieldClassName) throws Throwable {
 		/* Generate field class - value type with ImplicitCreation attribute, no flags */
 		byte[] fieldClassBytes = generateClass(fieldClassName, ACC_FINAL + ValhallaUtils.ACC_VALUE_TYPE,
-			new Attribute[] {new ImplicitCreationAttribute(0)});
+			new Attribute[] {new ValhallaUtils.ImplicitCreationAttribute(0)});
 		Class<?> fieldClass = generator.defineClass(fieldClassName, fieldClassBytes, 0, fieldClassBytes.length);
 
 		/* Generate class with field with NullRestricted attribute */
 		byte[] classBytes = generateIdentityClassWithField(className, isStatic ? ACC_STATIC : 0,
-			"field", fieldClass.descriptorString(), new Attribute[]{new NullRestrictedAttribute()});
+			"field", fieldClass.descriptorString(), new Attribute[]{new ValhallaUtils.NullRestrictedAttribute()});
 		return generator.defineClass(className, classBytes, 0, classBytes.length);
 	}
 
@@ -197,7 +191,7 @@ public class ValhallaAttributeGenerator extends ClassLoader {
 
 		/* Generate field class - value class with ImplicitCreation attribute and ACC_DEFAULT flag set.  */
 		byte[] fieldClassBytes = generateClass(fieldClassName, ACC_PUBLIC + ACC_FINAL + ValhallaUtils.ACC_VALUE_TYPE,
-			new Attribute[] {new ImplicitCreationAttribute(ValhallaUtils.ACC_DEFAULT)});
+			new Attribute[] {new ValhallaUtils.ImplicitCreationAttribute(ValhallaUtils.ACC_DEFAULT)});
 		Class<?> fieldClass = generator.defineClass(fieldClassName, fieldClassBytes, 0, fieldClassBytes.length);
 
 		ClassWriter classWriter = new ClassWriter(0);
@@ -205,7 +199,7 @@ public class ValhallaAttributeGenerator extends ClassLoader {
 
 		/* instance field of previously generated field class with NullRestrictd attribute */
 		FieldVisitor fieldVisitor = classWriter.visitField(ACC_PUBLIC, fieldName, fieldClass.descriptorString(), null, null);
-		fieldVisitor.visitAttribute(new NullRestrictedAttribute());
+		fieldVisitor.visitAttribute(new ValhallaUtils.NullRestrictedAttribute());
 
 		MethodVisitor mvInit = classWriter.visitMethod(ACC_PUBLIC, "<init>", "()V", null, null);
 		mvInit.visitCode();
@@ -268,94 +262,5 @@ public class ValhallaAttributeGenerator extends ClassLoader {
 
 		classWriter.visitEnd();
 		return classWriter.toByteArray();
-	}
-
-	final static class PreloadAttribute extends Attribute {
-		private final String[] classes;
-
-		public PreloadAttribute(String[] classes) {
-			super("Preload");
-			this.classes = classes;
-		}
-
-		public Label[] getLabels() {
-			return null;
-		}
-
-		public boolean isCodeAttribute() {
-			return false;
-		}
-
-		public boolean isUnknown() {
-			return true;
-		}
-
-		public ByteVector write(ClassWriter cw, byte[] code, int len, int maxStack, int maxLocals) {
-			ByteVector b = new ByteVector();
-
-			b.putShort(classes.length);
-
-			int cpIndex;
-			for (int i = 0; i < classes.length; i++) {
-				cpIndex = cw.newClass(classes[i].replace('.', '/'));
-				b.putShort(cpIndex);
-			}
-			return b;
-		}
-	}
-
-	final static class ImplicitCreationAttribute extends Attribute {
-		private final int flags;
-
-		public ImplicitCreationAttribute() {
-			super("ImplicitCreation");
-			/* this is the default flag generated by the compiler for the implicit modifier. */
-			this.flags = ValhallaUtils.ACC_DEFAULT;
-		}
-
-		public ImplicitCreationAttribute(int flags) {
-			super("ImplicitCreation");
-			this.flags = flags;
-		}
-
-		public Label[] getLabels() {
-			return null;
-		}
-
-		public boolean isCodeAttribute() {
-			return false;
-		}
-
-		public boolean isUnknown() {
-			return true;
-		}
-
-		public ByteVector write(ClassWriter cw, byte[] code, int len, int maxStack, int maxLocals) {
-			ByteVector b = new ByteVector();
-			b.putShort(flags);
-			return b;
-		}
-	}
-
-	final static class NullRestrictedAttribute extends Attribute {
-		public NullRestrictedAttribute() {
-			super("NullRestricted");
-		}
-
-		public Label[] getLabels() {
-			return null;
-		}
-
-		public boolean isCodeAttribute() {
-			return false;
-		}
-
-		public boolean isUnknown() {
-			return true;
-		}
-
-		public ByteVector write(ClassWriter cw, byte[] code, int len, int maxStack, int maxLocals) {
-			return new ByteVector();
-		}
 	}
 }
