+++ test/functional/Valhalla/src/org/openj9/test/lworld/ValhallaUtils.java
@@ -21,6 +21,8 @@
  *******************************************************************************/
 package org.openj9.test.lworld;
 
+import org.objectweb.asm.*;
+
 public class ValhallaUtils {
 	/**
 	 * Currently value type is built on JDK22, so use java file major version 66 for now.
@@ -40,4 +42,92 @@ public class ValhallaUtils {
 	static final int ACC_DEFAULT = 0x1;
 	static final int ACC_NON_ATOMIC = 0x2;
 
+	final static class PreloadAttribute extends Attribute {
+		private final String[] classes;
+
+		public PreloadAttribute(String[] classes) {
+			super("Preload");
+			this.classes = classes;
+		}
+
+		public Label[] getLabels() {
+			return null;
+		}
+
+		public boolean isCodeAttribute() {
+			return false;
+		}
+
+		public boolean isUnknown() {
+			return true;
+		}
+
+		public ByteVector write(ClassWriter cw, byte[] code, int len, int maxStack, int maxLocals) {
+			ByteVector b = new ByteVector();
+
+			b.putShort(classes.length);
+
+			int cpIndex;
+			for (int i = 0; i < classes.length; i++) {
+				cpIndex = cw.newClass(classes[i].replace('.', '/'));
+				b.putShort(cpIndex);
+			}
+			return b;
+		}
+	}
+
+	final static class ImplicitCreationAttribute extends Attribute {
+		private final int flags;
+
+		public ImplicitCreationAttribute() {
+			super("ImplicitCreation");
+			/* this is the default flag generated by the compiler for the implicit modifier. */
+			this.flags = ValhallaUtils.ACC_DEFAULT;
+		}
+
+		public ImplicitCreationAttribute(int flags) {
+			super("ImplicitCreation");
+			this.flags = flags;
+		}
+
+		public Label[] getLabels() {
+			return null;
+		}
+
+		public boolean isCodeAttribute() {
+			return false;
+		}
+
+		public boolean isUnknown() {
+			return true;
+		}
+
+		public ByteVector write(ClassWriter cw, byte[] code, int len, int maxStack, int maxLocals) {
+			ByteVector b = new ByteVector();
+			b.putShort(flags);
+			return b;
+		}
+	}
+
+	final static class NullRestrictedAttribute extends Attribute {
+		public NullRestrictedAttribute() {
+			super("NullRestricted");
+		}
+
+		public Label[] getLabels() {
+			return null;
+		}
+
+		public boolean isCodeAttribute() {
+			return false;
+		}
+
+		public boolean isUnknown() {
+			return true;
+		}
+
+		public ByteVector write(ClassWriter cw, byte[] code, int len, int maxStack, int maxLocals) {
+			return new ByteVector();
+		}
+	}
 }
